third_party/WebKit/LayoutTests/http/tests/flags/domstring.html

<!DOCTYPE html>
<title>Origin Flags API: Resources DOMString edge cases</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

function code_points(s) {
  return [...s]
    .map(c => '0x' + c.charCodeAt(0).toString(16).toUpperCase())
    .join(' ');
}

[
  '', // Empty strings
  'abc\x00def', // Embedded NUL
  '\uD800', // Unpaired low surrogage
  '\uDC00', // Unpaired high surrogage
  '\uDC00\uD800', // Swapped surrogate pair
  '\uFFFF' // Non-character
].forEach(string => {
  promise_test(async t => {
    const flag = await requestFlag(string, 'shared')
    assert_array_equals(flag.scope, [string],
                        'Requested name matches granted name');
  }, 'DOMString: ' + code_points(string));
});

promise_test(async t => {
  // '\uD800' treated as a USVString would become '\uFFFD'.
  let unblock;
  const blocked = new Promise(r => {unblock = r;});

  const flag = await requestFlag('\uD800', 'exclusive');
  assert_array_equals(flag.scope, ['\uD800']);
  flag.waitUntil(blocked);

  // Previous flag being held should not block this one
  // with a distinct DOMString.
  const flag2 = await requestFlag('\uFFFD', 'exclusive');
  assert_array_equals(flag2.scope, ['\uFFFD']);

  // If we did not time out, this passed.
  unblock();
}, 'Resource names that are not valid UTF-16 are not mangled');

</script>
